Re-introduce DID management methods
