+++ src/did/methods/index.js
